Move authenticator to api v1 path
