+++ app/controllers/api/v1/api_controller.rb
@@ -1,6 +1,8 @@
 module Api
   module V1
     class ApiController < ApplicationController
+      include Concerns::Authenticator
+
       before_action :auth_with_token!
     end
   end
