+++ spec/controllers/api/v1/concerns/authenticator_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
 
 class Authentication < ActionController::Base
-  include Concerns::Authenticable
+  include Api::V1::Concerns::Authenticator
 end
 
-describe Concerns::Authenticable do
+describe Api::V1::Concerns::Authenticator do
   let(:user) { FactoryGirl.create(:user) }
   let(:authentication) { Authentication.new }
   subject { authentication }
